+++ src/MachinaGrasshopper/Bridge/ActionIssued.cs
@@ -15,41 +15,39 @@ using MachinaGrasshopper.GH_Utils;
 
 namespace MachinaGrasshopper.Bridge
 {
-    //   █████╗  ██████╗████████╗██╗ ██████╗ ███╗   ██╗                   
-    //  ██╔══██╗██╔════╝╚══██╔══╝██║██╔═══██╗████╗  ██║                   
-    //  ███████║██║        ██║   ██║██║   ██║██╔██╗ ██║                   
-    //  ██╔══██║██║        ██║   ██║██║   ██║██║╚██╗██║                   
-    //  ██║  ██║╚██████╗   ██║   ██║╚██████╔╝██║ ╚████║                   
-    //  ╚═╝  ╚═╝ ╚═════╝   ╚═╝   ╚═╝ ╚═════╝ ╚═╝  ╚═══╝                   
-    //                                                                    
-    //  ███████╗██╗  ██╗███████╗ ██████╗██╗   ██╗████████╗███████╗██████╗ 
-    //  ██╔════╝╚██╗██╔╝██╔════╝██╔════╝██║   ██║╚══██╔══╝██╔════╝██╔══██╗
-    //  █████╗   ╚███╔╝ █████╗  ██║     ██║   ██║   ██║   █████╗  ██║  ██║
-    //  ██╔══╝   ██╔██╗ ██╔══╝  ██║     ██║   ██║   ██║   ██╔══╝  ██║  ██║
-    //  ███████╗██╔╝ ██╗███████╗╚██████╗╚██████╔╝   ██║   ███████╗██████╔╝
-    //  ╚══════╝╚═╝  ╚═╝╚══════╝ ╚═════╝ ╚═════╝    ╚═╝   ╚══════╝╚═════╝ 
-    //                                                                    
-    public class ActionExecuted : GH_Component
+    //   █████╗  ██████╗████████╗██╗ ██████╗ ███╗   ██╗
+    //  ██╔══██╗██╔════╝╚══██╔══╝██║██╔═══██╗████╗  ██║
+    //  ███████║██║        ██║   ██║██║   ██║██╔██╗ ██║
+    //  ██╔══██║██║        ██║   ██║██║   ██║██║╚██╗██║
+    //  ██║  ██║╚██████╗   ██║   ██║╚██████╔╝██║ ╚████║
+    //  ╚═╝  ╚═╝ ╚═════╝   ╚═╝   ╚═╝ ╚═════╝ ╚═╝  ╚═══╝
+    //                                                 
+    //  ██╗███████╗███████╗██╗   ██╗███████╗██████╗    
+    //  ██║██╔════╝██╔════╝██║   ██║██╔════╝██╔══██╗   
+    //  ██║███████╗███████╗██║   ██║█████╗  ██║  ██║   
+    //  ██║╚════██║╚════██║██║   ██║██╔══╝  ██║  ██║   
+    //  ██║███████║███████║╚██████╔╝███████╗██████╔╝   
+    //  ╚═╝╚══════╝╚══════╝ ╚═════╝ ╚══════╝╚═════╝    
+    //                                                 
+    public class ActionIssued : GH_Component
     {
         // For new events, all outputs will be updated, even if some of them have the same value (like position might be repeated on a Wait action...).
         private bool _updateOutputs;
-        private const string EVENT_NAME = "action-executed";
+        private const string EVENT_NAME = "action-issued";
 
         // Outputs
         private int _prevId, _id;
         private string _instruction;
-        private int _pendingExecutionOnDevice;
-        private int _pendingExecutionTotal;
         private Plane _tcp;
         private double?[] _axes;
         private double?[] _externalAxes;
 
         private JavaScriptSerializer ser;
 
-        public ActionExecuted() : base(
-            "ActionExecuted",
-            "ActionExecuted",
-            "Will update every time an Action has been successfully executed by the robot.",
+        public ActionIssued() : base(
+            "ActionIssued",
+            "ActionIssued",
+            "Will update every time an Action has been successfully issued and is pending release to the device.",
             "Machina",
             "Bridge")
         {
@@ -58,8 +56,8 @@ namespace MachinaGrasshopper.Bridge
         }
 
         public override GH_Exposure Exposure => GH_Exposure.secondary;
-        public override Guid ComponentGuid => new Guid("6aca9a1e-cdcf-435a-a627-7d8dda85ae6c");
-        protected override System.Drawing.Bitmap Icon => Properties.Resources.Bridge_ActionExecuted;
+        public override Guid ComponentGuid => new Guid("81f80c94-2ef4-4620-8248-227a1b07c3ad");
+        protected override System.Drawing.Bitmap Icon => Properties.Resources.Bridge_ActionIssued;
 
         protected override void RegisterInputParams(GH_Component.GH_InputParamManager pManager)
         {
@@ -68,14 +66,11 @@ namespace MachinaGrasshopper.Bridge
 
         protected override void RegisterOutputParams(GH_Component.GH_OutputParamManager pManager)
         {
-            pManager.AddTextParameter("LastAction", "last", "Last Action that was successfully executed by the robot.", GH_ParamAccess.item);
+            pManager.AddTextParameter("LastAction", "last", "Last Action that was successfully issued to the robot.", GH_ParamAccess.item);
 
             pManager.AddPlaneParameter("ActionTCP", "tcp", "Last known TCP position for this Action.", GH_ParamAccess.item);
             pManager.AddNumberParameter("ActionAxes", "axes", "Last known axes for this Action.", GH_ParamAccess.list);
             pManager.AddNumberParameter("ActionExternalAxes", "extax", "Last known external axes for this Action.", GH_ParamAccess.list);
-
-            pManager.AddNumberParameter("PendingActions", "pendTot", "How many Actions are left in the queue to be executed?", GH_ParamAccess.item);
-            pManager.AddNumberParameter("PendingActionsOnDevice", "pendDev", "How many Actions are left on the device to be executed? This only accounts for the ones that have already been released to it.", GH_ParamAccess.item);
         }
 
         protected override void ExpireDownStreamObjects()
@@ -104,8 +99,6 @@ namespace MachinaGrasshopper.Bridge
             DA.SetData(1, _tcp);
             DA.SetDataList(2, _axes);
             DA.SetDataList(3, _externalAxes);
-            DA.SetData(4, _pendingExecutionTotal);
-            DA.SetData(5, _pendingExecutionOnDevice);
 
             // If on second solution, stop checking.
             if (_updateOutputs)
@@ -180,9 +173,6 @@ namespace MachinaGrasshopper.Bridge
 
             _axes = Helpers.NullableDoublesFromObjects(json["axes"]);
             _externalAxes = Helpers.NullableDoublesFromObjects(json["extax"]);
-
-            _pendingExecutionOnDevice = json["pendDev"];
-            _pendingExecutionTotal = json["pendTot"];
         }
     }
 }