+++ src/MachinaGrasshopper/GH_Utils/GH_Utils.cs
@@ -4,6 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
+using Rhino.Geometry;
+using Grasshopper.Kernel;
+using Grasshopper.Kernel.Data;
+using Grasshopper.Kernel.Types;
+
+
 namespace MachinaGrasshopper.GH_Utils
 {
     /// <summary>
@@ -85,5 +91,54 @@ namespace MachinaGrasshopper.GH_Utils
 
             return d;
         }
+
+
+
+
+        internal static bool AreSimilar(Point3d a, Point3d b, double epsilon)
+        {
+            return Math.Abs(a.X - b.X) < epsilon && Math.Abs(a.Y - b.Y) < epsilon && Math.Abs(a.Z - b.Z) < epsilon;
+        }
+
+        internal static bool AreSimilar(Vector3d a, Vector3d b, double epsilon)
+        {
+            return Math.Abs(a.X - b.X) < epsilon && Math.Abs(a.Y - b.Y) < epsilon && Math.Abs(a.Z - b.Z) < epsilon;
+        }
+
+        internal static bool AreSimilar(Line a, Line b, double epsilon)
+        {
+            return AreSimilar(a.From, b.From, epsilon) && AreSimilar(a.To, b.To, epsilon);
+        }
+
+        internal static bool AreSimilar(Plane a, Plane b, double epsilon)
+        {
+            return AreSimilar(a.Origin, b.Origin, epsilon) && AreSimilar(a.XAxis, b.XAxis, epsilon) && AreSimilar(a.YAxis, b.YAxis, epsilon);
+        }
+
+        internal static bool AreSimilar(double?[] a, double?[] b, double epsilon)
+        {
+            if (a.Length != b.Length) return false;
+
+            double diff;
+            for (int i = 0; i < a.Length; i++)
+            {
+                try
+                {
+                    if (a[i] == null && b[i] == null) continue;
+                    diff = (double)(a[i] - b[i]);
+                    if (Math.Abs(diff) > epsilon)
+                        return false;
+                }
+                catch
+                {
+                    // Something went wrong with casting...
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
     }
+    
 }