+++ src/MachinaGrasshopper/Properties/Resources.Designer.cs
@@ -413,6 +413,16 @@ namespace MachinaGrasshopper.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap Robot_Logger {
+            get {
+                object obj = ResourceManager.GetObject("Robot_Logger", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap Robot_Version {
             get {
                 object obj = ResourceManager.GetObject("Robot_Version", resourceCulture);